// Angular 13 NGRX/src/app/features/user/enums/user-list.enum.ts
export enum UserListColumn {
      FirstName = 'firstName',
      Age = 'age',
      PhoneNumber = 'phoneNumber',
      Email = 'email',
      Address = 'address',
      Action = 'action'
}
 
export enum UserListHeader {
      NameHeader = 'Name',
      AgeHeader = 'Age',
      PhoneNumberHeader = 'Phone Number',
      EmailHeader = 'Email',
      AddressHeader = 'Address',
      ActionHeader = 'Action'
}

export enum UserPageTitle {
      UsersList = 'Users List',
      UserForm = 'User Registration Form'
}